[up] * : try to remove valgrind error

+++ include/server.h
@@ -93,7 +93,7 @@ void free_data_send(data_send_t *data_send);
 
 void init_data_struct(server_t *info);
 
-struct timespec set_timespec(int time, int freq);
+struct timespec set_timespec(long long int time, long long int freq);
 struct timespec sub_timespec(struct timespec ts1, struct timespec ts2);
 void select_interupt(server_t *info);
 void get_shortest_time(server_t *info);
